Move functions to new directory tree
