+++ pallets/asset/src/lib.rs
@@ -81,117 +81,42 @@
 pub mod benchmarking;
 pub mod checkpoint;
 
-#[cfg(feature = "std")]
-use sp_runtime::{Deserialize, Serialize};
+mod base_functions;
+mod base_validations_writes;
+mod error;
+mod rpc_functions;
+mod trait_impl;
+mod types;
 
-use arrayvec::ArrayVec;
 use codec::{Decode, Encode};
-use core::mem;
-use core::result::Result as StdResult;
-use currency::*;
-use frame_support::dispatch::{DispatchError, DispatchResult, Weight};
+use frame_support::dispatch::{DispatchResult, Weight};
 use frame_support::traits::{Get, PalletInfoAccess};
 use frame_support::BoundedBTreeSet;
-use frame_support::{decl_error, decl_module, decl_storage, ensure, fail};
-use frame_system::ensure_root;
-use scale_info::TypeInfo;
-use sp_runtime::traits::Zero;
-use sp_std::collections::btree_set::BTreeSet;
-use sp_std::{convert::TryFrom, prelude::*};
+use frame_support::{decl_module, decl_storage};
+use sp_std::prelude::*;
 
-use pallet_base::{
-    ensure_opt_string_limited, ensure_string_limited, try_next_pre, Error::CounterOverflow,
-};
-use pallet_identity::PermissionedCallOriginData;
-use polymesh_common_utilities::asset::AssetFnTrait;
-use polymesh_common_utilities::compliance_manager::ComplianceFnConfig;
-use polymesh_common_utilities::constants::*;
-use polymesh_common_utilities::protocol_fee::{ChargeProtocolFee, ProtocolOp};
 pub use polymesh_common_utilities::traits::asset::{Config, Event, RawEvent, WeightInfo};
-use polymesh_common_utilities::traits::nft::NFTTrait;
-use polymesh_common_utilities::with_transaction;
-
-use polymesh_primitives::agent::AgentGroup;
-use polymesh_primitives::asset::{
-    AssetName, AssetType, CheckpointId, CustomAssetTypeId, FundingRoundName,
-    GranularCanTransferResult,
-};
+use polymesh_primitives::asset::{AssetName, AssetType, CustomAssetTypeId, FundingRoundName};
 use polymesh_primitives::asset_metadata::{
     AssetMetadataGlobalKey, AssetMetadataKey, AssetMetadataLocalKey, AssetMetadataName,
     AssetMetadataSpec, AssetMetadataValue, AssetMetadataValueDetail,
 };
-use polymesh_primitives::settlement::InstructionId;
-use polymesh_primitives::transfer_compliance::TransferConditionResult;
 use polymesh_primitives::{
-    extract_auth, storage_migration_ver, AssetIdentifier, Balance, Document, DocumentId,
-    IdentityId, Memo, PortfolioId, PortfolioKind, PortfolioUpdateReason, SecondaryKey, Ticker,
-    WeightMeter,
+    storage_migration_ver, AssetIdentifier, Balance, Document, DocumentId, IdentityId, PortfolioId,
+    PortfolioKind, Ticker, WeightMeter,
+};
+
+pub use error::Error;
+pub use types::{
+    AssetOwnershipRelation, SecurityToken, TickerRegistration, TickerRegistrationConfig,
+    TickerRegistrationStatus,
 };
 
 type Checkpoint<T> = checkpoint::Module<T>;
 type ExternalAgents<T> = pallet_external_agents::Module<T>;
+type Identity<T> = pallet_identity::Module<T>;
 type Portfolio<T> = pallet_portfolio::Module<T>;
 type Statistics<T> = pallet_statistics::Module<T>;
-
-/// Ownership status of a ticker/token.
-#[derive(Encode, Decode, TypeInfo, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
-pub enum AssetOwnershipRelation {
-    NotOwned,
-    TickerOwned,
-    AssetOwned,
-}
-
-impl Default for AssetOwnershipRelation {
-    fn default() -> Self {
-        Self::NotOwned
-    }
-}
-
-/// struct to store the token details.
-#[derive(Encode, Decode, TypeInfo, Default, Clone, PartialEq, Debug)]
-pub struct SecurityToken {
-    pub total_supply: Balance,
-    pub owner_did: IdentityId,
-    pub divisible: bool,
-    pub asset_type: AssetType,
-}
-
-/// struct to store the ticker registration details.
-#[derive(Encode, Decode, TypeInfo, Clone, Default, PartialEq, Debug)]
-pub struct TickerRegistration<U> {
-    pub owner: IdentityId,
-    pub expiry: Option<U>,
-}
-
-/// struct to store the ticker registration config.
-#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
-#[derive(Encode, Decode, TypeInfo, Clone, Default, PartialEq, Debug)]
-pub struct TickerRegistrationConfig<U> {
-    pub max_ticker_length: u8,
-    pub registration_length: Option<U>,
-}
-
-/// Enum that represents the current status of a ticker.
-#[derive(Encode, Decode, Clone, Eq, PartialEq, Debug)]
-pub enum TickerRegistrationStatus {
-    RegisteredByOther,
-    Available,
-    RegisteredByDid,
-}
-
-/// Enum that uses as the return type for the restriction verification.
-#[derive(Encode, Decode, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
-pub enum RestrictionResult {
-    Valid,
-    Invalid,
-    ForceValid,
-}
-
-impl Default for RestrictionResult {
-    fn default() -> Self {
-        RestrictionResult::Invalid
-    }
-}
 
 storage_migration_ver!(3);
 
@@ -312,9 +237,6 @@ decl_storage! {
     }
 }
 
-type Identity<T> = pallet_identity::Module<T>;
-
-// Public interface for this runtime module.
 decl_module! {
     pub struct Module<T: Config> for enum Call where origin: T::RuntimeOrigin {
 
@@ -446,7 +368,7 @@ decl_module! {
         /// * Asset
         #[weight = <T as Config>::WeightInfo::freeze()]
         pub fn freeze(origin, ticker: Ticker) -> DispatchResult {
-            Self::set_freeze(origin, ticker, true)
+            Self::base_set_freeze(origin, ticker, true)
         }
 
         /// Unfreezes transfers of a given token.
@@ -462,7 +384,7 @@ decl_module! {
         /// * Asset
         #[weight = <T as Config>::WeightInfo::unfreeze()]
         pub fn unfreeze(origin, ticker: Ticker) -> DispatchResult {
-            Self::set_freeze(origin, ticker, false)
+            Self::base_set_freeze(origin, ticker, false)
         }
 
         /// Renames a given token.
@@ -935,1761 +857,5 @@ decl_module! {
         ) {
             Self::base_remove_mandatory_mediators(origin, ticker, mediators)?;
         }
-    }
-}
-
-decl_error! {
-    pub enum Error for Module<T: Config> {
-        /// The user is not authorized.
-        Unauthorized,
-        /// The token has already been created.
-        AssetAlreadyCreated,
-        /// The ticker length is over the limit.
-        TickerTooLong,
-        /// The ticker has non-alphanumeric parts.
-        TickerNotAlphanumeric,
-        /// The ticker is already registered to someone else.
-        TickerAlreadyRegistered,
-        /// The total supply is above the limit.
-        TotalSupplyAboveLimit,
-        /// No such token.
-        NoSuchAsset,
-        /// The token is already frozen.
-        AlreadyFrozen,
-        /// Not an owner of the token on Ethereum.
-        NotAnOwner,
-        /// An overflow while calculating the balance.
-        BalanceOverflow,
-        /// An overflow while calculating the total supply.
-        TotalSupplyOverflow,
-        /// An invalid granularity.
-        InvalidGranularity,
-        /// The asset must be frozen.
-        NotFrozen,
-        /// Transfer validation check failed.
-        InvalidTransfer,
-        /// The sender balance is not sufficient.
-        InsufficientBalance,
-        /// The token is already divisible.
-        AssetAlreadyDivisible,
-        /// An invalid Ethereum `EcdsaSignature`.
-        InvalidEthereumSignature,
-        /// Registration of ticker has expired.
-        TickerRegistrationExpired,
-        /// Transfers to self are not allowed
-        SenderSameAsReceiver,
-        /// The given Document does not exist.
-        NoSuchDoc,
-        /// Maximum length of asset name has been exceeded.
-        MaxLengthOfAssetNameExceeded,
-        /// Maximum length of the funding round name has been exceeded.
-        FundingRoundNameMaxLengthExceeded,
-        /// Some `AssetIdentifier` was invalid.
-        InvalidAssetIdentifier,
-        /// Investor Uniqueness claims are not allowed for this asset.
-        InvestorUniquenessClaimNotAllowed,
-        /// Invalid `CustomAssetTypeId`.
-        InvalidCustomAssetTypeId,
-        /// Maximum length of the asset metadata type name has been exceeded.
-        AssetMetadataNameMaxLengthExceeded,
-        /// Maximum length of the asset metadata value has been exceeded.
-        AssetMetadataValueMaxLengthExceeded,
-        /// Maximum length of the asset metadata type definition has been exceeded.
-        AssetMetadataTypeDefMaxLengthExceeded,
-        /// Asset Metadata key is missing.
-        AssetMetadataKeyIsMissing,
-        /// Asset Metadata value is locked.
-        AssetMetadataValueIsLocked,
-        /// Asset Metadata Local type already exists for asset.
-        AssetMetadataLocalKeyAlreadyExists,
-        /// Asset Metadata Global type already exists.
-        AssetMetadataGlobalKeyAlreadyExists,
-        /// Tickers should start with at least one valid byte.
-        TickerFirstByteNotValid,
-        /// Attempt to call an extrinsic that is only permitted for fungible tokens.
-        UnexpectedNonFungibleToken,
-        /// Attempt to update the type of a non fungible token to a fungible token or the other way around.
-        IncompatibleAssetTypeUpdate,
-        /// Attempt to delete a key that is needed for an NFT collection.
-        AssetMetadataKeyBelongsToNFTCollection,
-        /// Attempt to lock a metadata value that is empty.
-        AssetMetadataValueIsEmpty,
-        /// Number of asset mediators would exceed the maximum allowed.
-        NumberOfAssetMediatorsExceeded,
-        /// Invalid ticker character - valid set: A`..`Z` `0`..`9` `_` `-` `.` `/`.
-        InvalidTickerCharacter
-    }
-}
-
-impl<T: Config> Module<T> {
-    /// Issues `amount` tokens for `ticker` into the caller's portfolio.
-    fn base_issue(
-        origin: T::RuntimeOrigin,
-        ticker: Ticker,
-        amount: Balance,
-        portfolio_kind: PortfolioKind,
-    ) -> DispatchResult {
-        let portfolio_id = Self::ensure_origin_ticker_and_portfolio_permissions(
-            origin,
-            ticker,
-            portfolio_kind,
-            false,
-        )?;
-        let mut weight_meter = WeightMeter::max_limit_no_minimum();
-        Self::_mint(
-            &ticker,
-            portfolio_id.did,
-            amount,
-            Some(ProtocolOp::AssetIssue),
-            &mut weight_meter,
-        )
-    }
-
-    /// Ensures that `origin` is a permissioned agent for `ticker`, that the portfolio is valid and that calller
-    /// has the access to the portfolio. If `ensure_custody` is `true`, also enforces the caller to have custody
-    /// of the portfolio.
-    pub fn ensure_origin_ticker_and_portfolio_permissions(
-        origin: T::RuntimeOrigin,
-        ticker: Ticker,
-        portfolio_kind: PortfolioKind,
-        ensure_custody: bool,
-    ) -> Result<PortfolioId, DispatchError> {
-        let origin_data = <ExternalAgents<T>>::ensure_agent_asset_perms(origin, ticker)?;
-        let portfolio_id = PortfolioId::new(origin_data.primary_did, portfolio_kind);
-        Portfolio::<T>::ensure_portfolio_validity(&portfolio_id)?;
-        if ensure_custody {
-            Portfolio::<T>::ensure_portfolio_custody(portfolio_id, origin_data.primary_did)?;
-        }
-        Portfolio::<T>::ensure_user_portfolio_permission(
-            origin_data.secondary_key.as_ref(),
-            portfolio_id,
-        )?;
-        Ok(portfolio_id)
-    }
-
-    /// Ensure that all `idents` are valid.
-    fn ensure_asset_idents_valid(idents: &[AssetIdentifier]) -> DispatchResult {
-        ensure!(
-            idents.iter().all(|i| i.is_valid()),
-            Error::<T>::InvalidAssetIdentifier
-        );
-        Ok(())
-    }
-
-    /// Ensure `AssetType` is valid.
-    /// This checks that the `AssetType::Custom(custom_type_id)` is valid.
-    fn ensure_asset_type_valid(asset_type: AssetType) -> DispatchResult {
-        if let AssetType::Custom(custom_type_id) = asset_type {
-            ensure!(
-                CustomTypes::contains_key(custom_type_id),
-                Error::<T>::InvalidCustomAssetTypeId
-            );
-        }
-        Ok(())
-    }
-
-    pub fn base_register_ticker(origin: T::RuntimeOrigin, ticker: Ticker) -> DispatchResult {
-        let to_did = Identity::<T>::ensure_perms(origin)?;
-        let expiry = Self::ticker_registration_checks(&ticker, to_did, false, || {
-            Self::ticker_registration_config()
-        })?;
-
-        T::ProtocolFee::charge_fee(ProtocolOp::AssetRegisterTicker)?;
-        Self::unverified_register_ticker(&ticker, to_did, expiry);
-
-        Ok(())
-    }
-
-    /// Update identitifiers of `ticker` as `did`.
-    ///
-    /// Does not verify that actor `did` is permissioned for this call or that `idents` are valid.
-    fn unverified_update_idents(did: IdentityId, ticker: Ticker, idents: Vec<AssetIdentifier>) {
-        Identifiers::insert(ticker, idents.clone());
-        Self::deposit_event(RawEvent::IdentifiersUpdated(did, ticker, idents));
-    }
-
-    /// Ensure that `did` is the owner of `ticker`.
-    pub fn ensure_owner(ticker: &Ticker, did: IdentityId) -> DispatchResult {
-        ensure!(Self::is_owner(ticker, did), Error::<T>::Unauthorized);
-        Ok(())
-    }
-
-    /// Ensure that `ticker` is a valid created asset.
-    fn ensure_asset_exists(ticker: &Ticker) -> DispatchResult {
-        ensure!(Tokens::contains_key(&ticker), Error::<T>::NoSuchAsset);
-        Ok(())
-    }
-
-    /// Ensure that the document `doc` exists for `ticker`.
-    pub fn ensure_doc_exists(ticker: &Ticker, doc: &DocumentId) -> DispatchResult {
-        ensure!(
-            AssetDocuments::contains_key(ticker, doc),
-            Error::<T>::NoSuchDoc
-        );
-        Ok(())
-    }
-
-    pub fn is_owner(ticker: &Ticker, did: IdentityId) -> bool {
-        match Self::asset_ownership_relation(&did, &ticker) {
-            AssetOwnershipRelation::AssetOwned | AssetOwnershipRelation::TickerOwned => true,
-            AssetOwnershipRelation::NotOwned => false,
-        }
-    }
-
-    fn maybe_ticker(ticker: &Ticker) -> Option<TickerRegistration<T::Moment>> {
-        <Tickers<T>>::get(ticker)
-    }
-
-    pub fn is_ticker_available(ticker: &Ticker) -> bool {
-        // Assumes uppercase ticker
-        if let Some(ticker) = Self::maybe_ticker(ticker) {
-            ticker
-                .expiry
-                .filter(|&e| <pallet_timestamp::Pallet<T>>::get() > e)
-                .is_some()
-        } else {
-            true
-        }
-    }
-
-    /// Returns `true` if the ticker exists, is owned by `did`, and ticker hasn't expired.
-    pub fn is_ticker_registry_valid(ticker: &Ticker, did: IdentityId) -> bool {
-        // Assumes uppercase ticker
-        if let Some(ticker) = Self::maybe_ticker(ticker) {
-            let now = <pallet_timestamp::Pallet<T>>::get();
-            ticker.owner == did && ticker.expiry.filter(|&e| now > e).is_none()
-        } else {
-            false
-        }
-    }
-
-    /// Returns:
-    /// - `RegisteredByOther` if ticker is registered to someone else.
-    /// - `Available` if ticker is available for registry.
-    /// - `RegisteredByDid` if ticker is already registered to provided did.
-    pub fn is_ticker_available_or_registered_to(
-        ticker: &Ticker,
-        did: IdentityId,
-    ) -> TickerRegistrationStatus {
-        // Assumes uppercase ticker
-        match Self::maybe_ticker(ticker) {
-            Some(TickerRegistration { expiry, owner }) => match expiry {
-                // Ticker registered to someone but expired and can be registered again.
-                Some(expiry) if <pallet_timestamp::Pallet<T>>::get() > expiry => {
-                    TickerRegistrationStatus::Available
-                }
-                // Ticker is already registered to provided did (may or may not expire in future).
-                _ if owner == did => TickerRegistrationStatus::RegisteredByDid,
-                // Ticker registered to someone else and hasn't expired.
-                _ => TickerRegistrationStatus::RegisteredByOther,
-            },
-            // Ticker not registered yet.
-            None => TickerRegistrationStatus::Available,
-        }
-    }
-
-    /// Returns `Ok` if the ticker contains only the following characters: `A`..`Z` `0`..`9` `_` `-` `.` `/`.
-    pub fn verify_ticker_characters(ticker: &Ticker) -> DispatchResult {
-        let ticker_bytes = ticker.as_ref();
-
-        // The first byte of the ticker cannot be NULL
-        if *ticker_bytes.first().unwrap_or(&0) == 0 {
-            return Err(Error::<T>::TickerFirstByteNotValid.into());
-        }
-
-        // Allows the following characters: `A`..`Z` `0`..`9` `_` `-` `.` `/`
-        let valid_characters = BTreeSet::from([b'_', b'-', b'.', b'/']);
-        for (byte_index, ticker_byte) in ticker_bytes.iter().enumerate() {
-            if !ticker_byte.is_ascii_uppercase()
-                && !ticker_byte.is_ascii_digit()
-                && !valid_characters.contains(ticker_byte)
-            {
-                if ticker_bytes[byte_index..].iter().all(|byte| *byte == 0) {
-                    return Ok(());
-                }
-
-                return Err(Error::<T>::InvalidTickerCharacter.into());
-            }
-        }
-        Ok(())
-    }
-
-    /// Before registering a ticker, do some checks, and return the expiry moment.
-    fn ticker_registration_checks(
-        ticker: &Ticker,
-        to_did: IdentityId,
-        no_re_register: bool,
-        config: impl FnOnce() -> TickerRegistrationConfig<T::Moment>,
-    ) -> Result<Option<T::Moment>, DispatchError> {
-        Self::verify_ticker_characters(&ticker)?;
-        Self::ensure_asset_fresh(&ticker)?;
-
-        let config = config();
-
-        // Ensure the ticker is not too long.
-        Self::ensure_ticker_length(&ticker, &config)?;
-
-        // Ensure that the ticker is not registered by someone else (or `to_did`, possibly).
-        if match Self::is_ticker_available_or_registered_to(&ticker, to_did) {
-            TickerRegistrationStatus::RegisteredByOther => true,
-            TickerRegistrationStatus::RegisteredByDid => no_re_register,
-            _ => false,
-        } {
-            fail!(Error::<T>::TickerAlreadyRegistered);
-        }
-
-        Ok(config
-            .registration_length
-            .map(|exp| <pallet_timestamp::Pallet<T>>::get() + exp))
-    }
-
-    /// Registers the given `ticker` to the `owner` identity with an optional expiry time.
-    ///
-    /// ## Expected constraints
-    /// - `owner` should be a valid IdentityId.
-    /// - `ticker` should be valid, please see `ticker_registration_checks`.
-    /// - `ticker` should be available or already registered by `owner`.
-    fn unverified_register_ticker(ticker: &Ticker, owner: IdentityId, expiry: Option<T::Moment>) {
-        if let Some(ticker_details) = Self::maybe_ticker(ticker) {
-            AssetOwnershipRelations::remove(ticker_details.owner, ticker);
-        }
-
-        let ticker_registration = TickerRegistration { owner, expiry };
-
-        // Store ticker registration details
-        <Tickers<T>>::insert(ticker, ticker_registration);
-        AssetOwnershipRelations::insert(owner, ticker, AssetOwnershipRelation::TickerOwned);
-
-        Self::deposit_event(RawEvent::TickerRegistered(owner, *ticker, expiry));
-    }
-
-    // Get the total supply of an asset `id`.
-    pub fn total_supply(ticker: Ticker) -> Balance {
-        Self::token_details(&ticker)
-            .map(|t| t.total_supply)
-            .unwrap_or_default()
-    }
-
-    pub fn get_balance_at(ticker: Ticker, did: IdentityId, at: CheckpointId) -> Balance {
-        <Checkpoint<T>>::balance_at(ticker, did, at)
-            .unwrap_or_else(|| Self::balance_of(&ticker, &did))
-    }
-
-    pub fn _is_valid_transfer(
-        ticker: &Ticker,
-        from_portfolio: PortfolioId,
-        to_portfolio: PortfolioId,
-        value: Balance,
-        weight_meter: &mut WeightMeter,
-    ) -> StdResult<u8, DispatchError> {
-        if Self::frozen(ticker) {
-            return Ok(ERC1400_TRANSFERS_HALTED);
-        }
-
-        if Self::portfolio_failure(&from_portfolio, &to_portfolio, ticker, value) {
-            return Ok(PORTFOLIO_FAILURE);
-        }
-
-        if Self::statistics_failures(
-            &from_portfolio.did,
-            &to_portfolio.did,
-            ticker,
-            value,
-            weight_meter,
-        ) {
-            return Ok(TRANSFER_MANAGER_FAILURE);
-        }
-
-        if !T::ComplianceManager::is_compliant(
-            ticker,
-            from_portfolio.did,
-            to_portfolio.did,
-            weight_meter,
-        )? {
-            return Ok(COMPLIANCE_MANAGER_FAILURE);
-        }
-
-        Ok(ERC1400_TRANSFER_SUCCESS)
-    }
-
-    // Transfers tokens from one identity to another
-    pub fn unsafe_transfer(
-        from_portfolio: PortfolioId,
-        to_portfolio: PortfolioId,
-        ticker: &Ticker,
-        value: Balance,
-        instruction_id: Option<InstructionId>,
-        instruction_memo: Option<Memo>,
-        caller_did: IdentityId,
-        weight_meter: &mut WeightMeter,
-    ) -> DispatchResult {
-        Self::ensure_granular(ticker, value)?;
-
-        let token = Self::token_details(ticker)?;
-        // Ensures the token is fungible
-        ensure!(
-            token.asset_type.is_fungible(),
-            Error::<T>::UnexpectedNonFungibleToken
-        );
-
-        ensure!(
-            from_portfolio.did != to_portfolio.did,
-            Error::<T>::SenderSameAsReceiver
-        );
-
-        let from_total_balance = Self::balance_of(ticker, from_portfolio.did);
-        ensure!(from_total_balance >= value, Error::<T>::InsufficientBalance);
-        let updated_from_total_balance = from_total_balance - value;
-
-        let to_total_balance = Self::balance_of(ticker, to_portfolio.did);
-        let updated_to_total_balance = to_total_balance
-            .checked_add(value)
-            .ok_or(Error::<T>::BalanceOverflow)?;
-
-        <Checkpoint<T>>::advance_update_balances(
-            ticker,
-            &[
-                (from_portfolio.did, from_total_balance),
-                (to_portfolio.did, to_total_balance),
-            ],
-        )?;
-
-        // reduce sender's balance
-        BalanceOf::insert(ticker, &from_portfolio.did, updated_from_total_balance);
-        // increase receiver's balance
-        BalanceOf::insert(ticker, &to_portfolio.did, updated_to_total_balance);
-        // transfer portfolio balances
-        Portfolio::<T>::unchecked_transfer_portfolio_balance(
-            &from_portfolio,
-            &to_portfolio,
-            ticker,
-            value,
-        );
-
-        // Update statistic info.
-        Statistics::<T>::update_asset_stats(
-            ticker,
-            Some(&from_portfolio.did),
-            Some(&to_portfolio.did),
-            Some(updated_from_total_balance),
-            Some(updated_to_total_balance),
-            value,
-            weight_meter,
-        )?;
-
-        Self::deposit_event(RawEvent::AssetBalanceUpdated(
-            caller_did,
-            *ticker,
-            value,
-            Some(from_portfolio),
-            Some(to_portfolio),
-            PortfolioUpdateReason::Transferred {
-                instruction_id,
-                instruction_memo,
-            },
-        ));
-        Ok(())
-    }
-
-    fn _mint(
-        ticker: &Ticker,
-        to_did: IdentityId,
-        value: Balance,
-        protocol_fee_data: Option<ProtocolOp>,
-        weight_meter: &mut WeightMeter,
-    ) -> DispatchResult {
-        Self::ensure_granular(ticker, value)?;
-        // Read the token details
-        let mut token = Self::token_details(ticker)?;
-        // Ensures the token is fungible
-        ensure!(
-            token.asset_type.is_fungible(),
-            Error::<T>::UnexpectedNonFungibleToken
-        );
-
-        // Prepare the updated total supply.
-        let updated_total_supply = token
-            .total_supply
-            .checked_add(value)
-            .ok_or(Error::<T>::TotalSupplyOverflow)?;
-        Self::ensure_within_max_supply(updated_total_supply)?;
-        // Increase receiver balance.
-        let current_to_balance = Self::balance_of(ticker, to_did);
-        // No check since the total balance is always <= the total supply. The
-        // total supply is already checked above.
-        let updated_to_balance = current_to_balance + value;
-        // No check since the default portfolio balance is always <= the total
-        // supply. The total supply is already checked above.
-        let updated_to_def_balance = Portfolio::<T>::portfolio_asset_balances(
-            PortfolioId::default_portfolio(to_did),
-            ticker,
-        ) + value;
-
-        // In transaction because we don't want fee to be charged if advancing fails.
-        with_transaction(|| {
-            // Charge the fee.
-            if let Some(op) = protocol_fee_data {
-                T::ProtocolFee::charge_fee(op)?;
-            }
-
-            // Advance checkpoint schedules and update last checkpoint.
-            <Checkpoint<T>>::advance_update_balances(ticker, &[(to_did, current_to_balance)])
-        })?;
-
-        // Increase total supply.
-        token.total_supply = updated_total_supply;
-        BalanceOf::insert(ticker, &to_did, updated_to_balance);
-        Portfolio::<T>::set_default_portfolio_balance(to_did, ticker, updated_to_def_balance);
-        Tokens::insert(ticker, token);
-
-        Statistics::<T>::update_asset_stats(
-            &ticker,
-            None,
-            Some(&to_did),
-            None,
-            Some(updated_to_balance),
-            value,
-            weight_meter,
-        )?;
-
-        let round = Self::funding_round(ticker);
-        let ticker_round = (*ticker, round.clone());
-        // No check since the issued balance is always <= the total
-        // supply. The total supply is already checked above.
-        let issued_in_this_round = Self::issued_in_funding_round(&ticker_round) + value;
-        IssuedInFundingRound::insert(&ticker_round, issued_in_this_round);
-
-        Self::deposit_event(RawEvent::AssetBalanceUpdated(
-            to_did,
-            *ticker,
-            value,
-            None,
-            Some(PortfolioId::default_portfolio(to_did)),
-            PortfolioUpdateReason::Issued {
-                funding_round_name: Some(round),
-            },
-        ));
-        Ok(())
-    }
-
-    fn ensure_granular(ticker: &Ticker, value: Balance) -> DispatchResult {
-        ensure!(
-            Self::check_granularity(&ticker, value),
-            Error::<T>::InvalidGranularity
-        );
-        Ok(())
-    }
-
-    fn check_granularity(ticker: &Ticker, value: Balance) -> bool {
-        Self::is_divisible(ticker) || Self::is_unit_multiple(value)
-    }
-
-    /// Is `value` a multiple of "one unit"?
-    fn is_unit_multiple(value: Balance) -> bool {
-        value % ONE_UNIT == 0
-    }
-
-    pub fn token_details(ticker: &Ticker) -> Result<SecurityToken, DispatchError> {
-        Ok(Tokens::try_get(ticker).or(Err(Error::<T>::NoSuchAsset))?)
-    }
-
-    pub fn is_divisible(ticker: &Ticker) -> bool {
-        Self::token_details(ticker)
-            .map(|t| t.divisible)
-            .unwrap_or_default()
-    }
-
-    /// Accepts and executes the ticker transfer.
-    fn base_accept_ticker_transfer(origin: T::RuntimeOrigin, auth_id: u64) -> DispatchResult {
-        let to = Identity::<T>::ensure_perms(origin)?;
-        <Identity<T>>::accept_auth_with(&to.into(), auth_id, |data, auth_by| {
-            let ticker = extract_auth!(data, TransferTicker(t));
-
-            Self::ensure_asset_fresh(&ticker)?;
-
-            let reg =
-                Self::ticker_registration(&ticker).ok_or(Error::<T>::TickerRegistrationExpired)?;
-            <Identity<T>>::ensure_auth_by(auth_by, reg.owner)?;
-
-            Self::transfer_ticker(reg, ticker, to);
-            Ok(())
-        })
-    }
-
-    /// Transfer the given `ticker`'s registration from `req.owner` to `to`.
-    fn transfer_ticker(mut reg: TickerRegistration<T::Moment>, ticker: Ticker, to: IdentityId) {
-        let from = reg.owner;
-        AssetOwnershipRelations::remove(from, ticker);
-        AssetOwnershipRelations::insert(to, ticker, AssetOwnershipRelation::TickerOwned);
-        reg.owner = to;
-        <Tickers<T>>::insert(&ticker, reg);
-        Self::deposit_event(RawEvent::TickerTransferred(to, ticker, from));
-    }
-
-    /// Accept and process a token ownership transfer.
-    fn base_accept_token_ownership_transfer(origin: T::RuntimeOrigin, id: u64) -> DispatchResult {
-        let to = Identity::<T>::ensure_perms(origin)?;
-        <Identity<T>>::accept_auth_with(&to.into(), id, |data, auth_by| {
-            let ticker = extract_auth!(data, TransferAssetOwnership(t));
-
-            // Get the token details and ensure it exists.
-            let mut token = Self::token_details(&ticker)?;
-
-            // Ensure the authorization was created by a permissioned agent.
-            <ExternalAgents<T>>::ensure_agent_permissioned(ticker, auth_by)?;
-
-            // Get the ticker registration and ensure it exists.
-            let mut reg = Self::ticker_registration(&ticker).ok_or(Error::<T>::NoSuchAsset)?;
-            let old_owner = reg.owner;
-            AssetOwnershipRelations::remove(old_owner, ticker);
-            AssetOwnershipRelations::insert(to, ticker, AssetOwnershipRelation::AssetOwned);
-            // Update ticker registration.
-            reg.owner = to;
-            <Tickers<T>>::insert(&ticker, reg);
-            // Update token details.
-            token.owner_did = to;
-            Tokens::insert(&ticker, token);
-            Self::deposit_event(RawEvent::AssetOwnershipTransferred(to, ticker, old_owner));
-            Ok(())
-        })
-    }
-
-    /// RPC: Function allows external users to know whether the transfer extrinsic
-    /// will be valid or not beforehand.
-    pub fn unsafe_can_transfer(
-        from_custodian: Option<IdentityId>,
-        from_portfolio: PortfolioId,
-        to_custodian: Option<IdentityId>,
-        to_portfolio: PortfolioId,
-        ticker: &Ticker,
-        value: Balance,
-        weight_meter: &mut WeightMeter,
-    ) -> StdResult<u8, &'static str> {
-        Ok(if Self::invalid_granularity(ticker, value) {
-            // Granularity check
-            INVALID_GRANULARITY
-        } else if Self::self_transfer(&from_portfolio, &to_portfolio) {
-            INVALID_RECEIVER_DID
-        } else if Self::invalid_cdd(from_portfolio.did) {
-            INVALID_SENDER_DID
-        } else if Self::custodian_error(
-            from_portfolio,
-            from_custodian.unwrap_or(from_portfolio.did),
-        ) {
-            CUSTODIAN_ERROR
-        } else if Self::invalid_cdd(to_portfolio.did) {
-            INVALID_RECEIVER_DID
-        } else if Self::custodian_error(to_portfolio, to_custodian.unwrap_or(to_portfolio.did)) {
-            CUSTODIAN_ERROR
-        } else if Self::insufficient_balance(&ticker, from_portfolio.did, value) {
-            ERC1400_INSUFFICIENT_BALANCE
-        } else if Self::portfolio_failure(&from_portfolio, &to_portfolio, ticker, value) {
-            PORTFOLIO_FAILURE
-        } else {
-            // Compliance manager & Smart Extension check
-            Self::_is_valid_transfer(&ticker, from_portfolio, to_portfolio, value, weight_meter)
-                .unwrap_or(ERC1400_TRANSFER_FAILURE)
-        })
-    }
-
-    /// Transfers an asset from one identity portfolio to another
-    pub fn base_transfer(
-        from_portfolio: PortfolioId,
-        to_portfolio: PortfolioId,
-        ticker: &Ticker,
-        value: Balance,
-        instruction_id: Option<InstructionId>,
-        instruction_memo: Option<Memo>,
-        caller_did: IdentityId,
-        weight_meter: &mut WeightMeter,
-    ) -> DispatchResult {
-        // NB: This function does not check if the sender/receiver have custodian permissions on the portfolios.
-        // The custodian permissions must be checked before this function is called.
-        // The only place this function is used right now is the settlement engine and the settlement engine
-        // checks custodial permissions when the instruction is authorized.
-
-        // Validate the transfer
-        let is_transfer_success =
-            Self::_is_valid_transfer(&ticker, from_portfolio, to_portfolio, value, weight_meter)?;
-
-        ensure!(
-            is_transfer_success == ERC1400_TRANSFER_SUCCESS,
-            Error::<T>::InvalidTransfer
-        );
-
-        Self::unsafe_transfer(
-            from_portfolio,
-            to_portfolio,
-            ticker,
-            value,
-            instruction_id,
-            instruction_memo,
-            caller_did,
-            weight_meter,
-        )?;
-
-        Ok(())
-    }
-
-    /// Performs necessary checks on parameters of `create_asset`.
-    fn ensure_create_asset_parameters(ticker: &Ticker) -> DispatchResult {
-        Self::ensure_asset_fresh(&ticker)?;
-        Self::ensure_ticker_length(&ticker, &Self::ticker_registration_config())
-    }
-
-    /// Ensure asset `ticker` doesn't exist yet.
-    fn ensure_asset_fresh(ticker: &Ticker) -> DispatchResult {
-        ensure!(
-            !Tokens::contains_key(ticker),
-            Error::<T>::AssetAlreadyCreated
-        );
-        Ok(())
-    }
-
-    /// Ensure `supply <= MAX_SUPPLY`.
-    fn ensure_within_max_supply(supply: Balance) -> DispatchResult {
-        ensure!(supply <= MAX_SUPPLY, Error::<T>::TotalSupplyAboveLimit);
-        Ok(())
-    }
-
-    /// Ensure ticker length is within limit per `config`.
-    fn ensure_ticker_length<U>(
-        ticker: &Ticker,
-        config: &TickerRegistrationConfig<U>,
-    ) -> DispatchResult {
-        ensure!(
-            ticker.len() <= usize::try_from(config.max_ticker_length).unwrap_or_default(),
-            Error::<T>::TickerTooLong
-        );
-        Ok(())
-    }
-
-    fn base_create_asset(
-        origin: T::RuntimeOrigin,
-        name: AssetName,
-        ticker: Ticker,
-        divisible: bool,
-        asset_type: AssetType,
-        identifiers: Vec<AssetIdentifier>,
-        funding_round: Option<FundingRoundName>,
-    ) -> Result<IdentityId, DispatchError> {
-        let PermissionedCallOriginData {
-            primary_did,
-            secondary_key,
-            ..
-        } = Identity::<T>::ensure_origin_call_permissions(origin)?;
-        Self::unsafe_create_asset(
-            primary_did,
-            secondary_key,
-            name,
-            ticker,
-            divisible,
-            asset_type,
-            identifiers,
-            funding_round,
-        )
-    }
-
-    fn unsafe_create_asset(
-        did: IdentityId,
-        secondary_key: Option<SecondaryKey<T::AccountId>>,
-        name: AssetName,
-        ticker: Ticker,
-        divisible: bool,
-        asset_type: AssetType,
-        identifiers: Vec<AssetIdentifier>,
-        funding_round: Option<FundingRoundName>,
-    ) -> Result<IdentityId, DispatchError> {
-        Self::ensure_asset_name_bounded(&name)?;
-        if let Some(fr) = &funding_round {
-            Self::ensure_funding_round_name_bounded(fr)?;
-        }
-        Self::ensure_asset_idents_valid(&identifiers)?;
-        Self::ensure_asset_type_valid(asset_type)?;
-
-        Self::ensure_create_asset_parameters(&ticker)?;
-
-        // Ensure its registered by DID or at least expired, thus available.
-        let available = match Self::is_ticker_available_or_registered_to(&ticker, did) {
-            TickerRegistrationStatus::RegisteredByOther => {
-                fail!(Error::<T>::TickerAlreadyRegistered)
-            }
-            TickerRegistrationStatus::RegisteredByDid => false,
-            TickerRegistrationStatus::Available => true,
-        };
-
-        // If `ticker` isn't registered, it will be, so ensure it is fully ascii.
-        if available {
-            Self::verify_ticker_characters(&ticker)?;
-        }
-
-        let token_did = Identity::<T>::get_token_did(&ticker)?;
-        // Ensure there's no pre-existing entry for the DID.
-        // This should never happen, but let's be defensive here.
-        Identity::<T>::ensure_no_id_record(token_did)?;
-
-        // Ensure that the caller has relevant portfolio permissions
-        let user_default_portfolio = PortfolioId::default_portfolio(did);
-        Portfolio::<T>::ensure_portfolio_custody_and_permission(
-            user_default_portfolio,
-            did,
-            secondary_key.as_ref(),
-        )?;
-
-        // Charge protocol fees.
-        T::ProtocolFee::charge_fees(&{
-            let mut fees = ArrayVec::<_, 2>::new();
-            if available {
-                fees.push(ProtocolOp::AssetRegisterTicker);
-                fees.push(ProtocolOp::AssetCreateAsset);
-            }
-            fees
-        })?;
-
-        //==========================================================================
-        // At this point all checks have been made; **only** storage changes follow!
-        //==========================================================================
-
-        Identity::<T>::commit_token_did(token_did, ticker);
-
-        // Register the ticker or finish its registration.
-        if available {
-            // Ticker not registered by anyone (or registry expired), so register.
-            Self::unverified_register_ticker(&ticker, did, None);
-        } else {
-            // Ticker already registered by the user.
-            <Tickers<T>>::mutate(&ticker, |tr| {
-                if let Some(tr) = tr {
-                    tr.expiry = None;
-                }
-            });
-        }
-
-        let token = SecurityToken {
-            total_supply: Zero::zero(),
-            owner_did: did,
-            divisible,
-            asset_type,
-        };
-        Tokens::insert(&ticker, token);
-        AssetNames::insert(&ticker, &name);
-        // NB - At the time of asset creation it is obvious that the asset issuer will not have an
-        // `InvestorUniqueness` claim. So we are skipping the scope claim based stats update as
-        // those data points will get added in to the system whenever the asset issuer
-        // has an InvestorUniqueness claim. This also applies when issuing assets.
-        AssetOwnershipRelations::insert(did, ticker, AssetOwnershipRelation::AssetOwned);
-        Self::deposit_event(RawEvent::AssetCreated(
-            did,
-            ticker,
-            divisible,
-            asset_type,
-            did,
-            name,
-            identifiers.clone(),
-            funding_round.clone(),
-        ));
-
-        // Add funding round name.
-        if let Some(funding_round) = funding_round {
-            FundingRound::insert(ticker, funding_round);
-        }
-
-        Self::unverified_update_idents(did, ticker, identifiers);
-
-        // Grant owner full agent permissions.
-        <ExternalAgents<T>>::unchecked_add_agent(ticker, did, AgentGroup::Full).unwrap();
-
-        Ok(did)
-    }
-
-    fn set_freeze(origin: T::RuntimeOrigin, ticker: Ticker, freeze: bool) -> DispatchResult {
-        let did = <ExternalAgents<T>>::ensure_perms(origin, ticker)?;
-        Self::ensure_asset_exists(&ticker)?;
-
-        let (event, error) = match freeze {
-            true => (
-                RawEvent::AssetFrozen(did, ticker),
-                Error::<T>::AlreadyFrozen,
-            ),
-            false => (RawEvent::AssetUnfrozen(did, ticker), Error::<T>::NotFrozen),
-        };
-
-        ensure!(Self::frozen(&ticker) != freeze, error);
-        Frozen::insert(&ticker, freeze);
-
-        Self::deposit_event(event);
-        Ok(())
-    }
-
-    fn base_rename_asset(
-        origin: T::RuntimeOrigin,
-        ticker: Ticker,
-        name: AssetName,
-    ) -> DispatchResult {
-        Self::ensure_asset_name_bounded(&name)?;
-        Self::ensure_asset_exists(&ticker)?;
-        let did = <ExternalAgents<T>>::ensure_perms(origin, ticker)?;
-
-        AssetNames::insert(&ticker, name.clone());
-        Self::deposit_event(RawEvent::AssetRenamed(did, ticker, name));
-        Ok(())
-    }
-
-    /// Ensure `name` is within the global limit for asset name lengths.
-    fn ensure_asset_name_bounded(name: &AssetName) -> DispatchResult {
-        ensure!(
-            name.len() as u32 <= T::AssetNameMaxLength::get(),
-            Error::<T>::MaxLengthOfAssetNameExceeded
-        );
-        Ok(())
-    }
-
-    fn base_redeem(
-        origin: T::RuntimeOrigin,
-        ticker: Ticker,
-        value: Balance,
-        portfolio_kind: PortfolioKind,
-        weight_meter: &mut WeightMeter,
-    ) -> DispatchResult {
-        let portfolio = Self::ensure_origin_ticker_and_portfolio_permissions(
-            origin,
-            ticker,
-            portfolio_kind,
-            true,
-        )?;
-
-        Self::ensure_granular(&ticker, value)?;
-
-        let mut token = Self::token_details(&ticker)?;
-        // Ensures the token is fungible
-        ensure!(
-            token.asset_type.is_fungible(),
-            Error::<T>::UnexpectedNonFungibleToken
-        );
-
-        // Reduce caller's portfolio balance. This makes sure that the caller has enough unlocked tokens.
-        // If `advance_update_balances` fails, `reduce_portfolio_balance` shouldn't modify storage.
-
-        with_transaction(|| {
-            Portfolio::<T>::reduce_portfolio_balance(&portfolio, &ticker, value)?;
-
-            // Try updating the total supply.
-            token.total_supply = token
-                .total_supply
-                .checked_sub(value)
-                .ok_or(Error::<T>::TotalSupplyOverflow)?;
-
-            <Checkpoint<T>>::advance_update_balances(
-                &ticker,
-                &[(portfolio.did, Self::balance_of(ticker, portfolio.did))],
-            )
-        })?;
-
-        let updated_balance = Self::balance_of(ticker, portfolio.did) - value;
-
-        // Update identity balances and total supply
-        BalanceOf::insert(ticker, &portfolio.did, updated_balance);
-        Tokens::insert(ticker, token);
-
-        // Update statistic info.
-        Statistics::<T>::update_asset_stats(
-            &ticker,
-            Some(&portfolio.did),
-            None,
-            Some(updated_balance),
-            None,
-            value,
-            weight_meter,
-        )?;
-
-        Self::deposit_event(RawEvent::AssetBalanceUpdated(
-            portfolio.did,
-            ticker,
-            value,
-            Some(portfolio),
-            None,
-            PortfolioUpdateReason::Redeemed,
-        ));
-        Ok(())
-    }
-
-    fn base_make_divisible(origin: T::RuntimeOrigin, ticker: Ticker) -> DispatchResult {
-        let did = <ExternalAgents<T>>::ensure_perms(origin, ticker)?;
-
-        Tokens::try_mutate(&ticker, |token| -> DispatchResult {
-            let token = token.as_mut().ok_or(Error::<T>::NoSuchAsset)?;
-            // Ensures the token is fungible
-            ensure!(
-                token.asset_type.is_fungible(),
-                Error::<T>::UnexpectedNonFungibleToken
-            );
-            ensure!(!token.divisible, Error::<T>::AssetAlreadyDivisible);
-            token.divisible = true;
-
-            Self::deposit_event(RawEvent::DivisibilityChanged(did, ticker, true));
-            Ok(())
-        })
-    }
-
-    fn base_add_documents(
-        origin: T::RuntimeOrigin,
-        docs: Vec<Document>,
-        ticker: Ticker,
-    ) -> DispatchResult {
-        let did = <ExternalAgents<T>>::ensure_perms(origin, ticker)?;
-
-        // Ensure strings are limited.
-        for doc in &docs {
-            ensure_string_limited::<T>(&doc.uri)?;
-            ensure_string_limited::<T>(&doc.name)?;
-            ensure_opt_string_limited::<T>(doc.doc_type.as_deref())?;
-        }
-
-        // Ensure we can advance documents ID sequence by `len`.
-        let pre = AssetDocumentsIdSequence::try_mutate(ticker, |id| {
-            id.0.checked_add(docs.len() as u32)
-                .ok_or(CounterOverflow::<T>)
-                .map(|new| mem::replace(id, DocumentId(new)))
-        })?;
-
-        // Charge fee.
-        T::ProtocolFee::batch_charge_fee(ProtocolOp::AssetAddDocuments, docs.len())?;
-
-        // Add the documents & emit events.
-        for (id, doc) in (pre.0..).map(DocumentId).zip(docs) {
-            AssetDocuments::insert(ticker, id, doc.clone());
-            Self::deposit_event(RawEvent::DocumentAdded(did, ticker, id, doc));
-        }
-        Ok(())
-    }
-
-    fn base_remove_documents(
-        origin: T::RuntimeOrigin,
-        ids: Vec<DocumentId>,
-        ticker: Ticker,
-    ) -> DispatchResult {
-        let did = <ExternalAgents<T>>::ensure_perms(origin, ticker)?;
-        for id in ids {
-            AssetDocuments::remove(ticker, id);
-            Self::deposit_event(RawEvent::DocumentRemoved(did, ticker, id));
-        }
-        Ok(())
-    }
-
-    fn base_set_funding_round(
-        origin: T::RuntimeOrigin,
-        ticker: Ticker,
-        name: FundingRoundName,
-    ) -> DispatchResult {
-        Self::ensure_funding_round_name_bounded(&name)?;
-        let did = <ExternalAgents<T>>::ensure_perms(origin, ticker)?;
-
-        FundingRound::insert(ticker, name.clone());
-        Self::deposit_event(RawEvent::FundingRoundSet(did, ticker, name));
-        Ok(())
-    }
-
-    /// Ensure `name` is within the global limit for asset name lengths.
-    fn ensure_funding_round_name_bounded(name: &FundingRoundName) -> DispatchResult {
-        ensure!(
-            name.len() as u32 <= T::FundingRoundNameMaxLength::get(),
-            Error::<T>::FundingRoundNameMaxLengthExceeded
-        );
-        Ok(())
-    }
-
-    fn base_update_identifiers(
-        origin: T::RuntimeOrigin,
-        ticker: Ticker,
-        identifiers: Vec<AssetIdentifier>,
-    ) -> DispatchResult {
-        let did = <ExternalAgents<T>>::ensure_perms(origin, ticker)?;
-        Self::ensure_asset_idents_valid(&identifiers)?;
-        Self::unverified_update_idents(did, ticker, identifiers);
-        Ok(())
-    }
-
-    fn is_asset_metadata_locked(ticker: Ticker, key: AssetMetadataKey) -> bool {
-        AssetMetadataValueDetails::<T>::get(ticker, key).map_or(false, |details| {
-            details.is_locked(<pallet_timestamp::Pallet<T>>::get())
-        })
-    }
-
-    pub fn check_asset_metadata_key_exists(ticker: &Ticker, key: &AssetMetadataKey) -> bool {
-        match key {
-            AssetMetadataKey::Global(key) => AssetMetadataGlobalKeyToName::contains_key(key),
-            AssetMetadataKey::Local(key) => AssetMetadataLocalKeyToName::contains_key(ticker, key),
-        }
-    }
-
-    /// Ensure asset metadata `value` is within the global limit.
-    fn ensure_asset_metadata_value_limited(value: &AssetMetadataValue) -> DispatchResult {
-        ensure!(
-            value.len() <= T::AssetMetadataValueMaxLength::get() as usize,
-            Error::<T>::AssetMetadataValueMaxLengthExceeded
-        );
-        Ok(())
-    }
-
-    /// Ensure asset metadata `name` is within the global limit.
-    fn ensure_asset_metadata_name_limited(name: &AssetMetadataName) -> DispatchResult {
-        ensure!(
-            name.len() <= T::AssetMetadataNameMaxLength::get() as usize,
-            Error::<T>::AssetMetadataNameMaxLengthExceeded
-        );
-        Ok(())
-    }
-
-    /// Ensure asset metadata `spec` is within the global limit.
-    fn ensure_asset_metadata_spec_limited(spec: &AssetMetadataSpec) -> DispatchResult {
-        ensure_opt_string_limited::<T>(spec.url.as_deref())?;
-        ensure_opt_string_limited::<T>(spec.description.as_deref())?;
-        if let Some(ref type_def) = spec.type_def {
-            ensure!(
-                type_def.len() <= T::AssetMetadataTypeDefMaxLength::get() as usize,
-                Error::<T>::AssetMetadataTypeDefMaxLengthExceeded
-            );
-        }
-        Ok(())
-    }
-
-    fn base_set_asset_metadata(
-        origin: T::RuntimeOrigin,
-        ticker: Ticker,
-        key: AssetMetadataKey,
-        value: AssetMetadataValue,
-        detail: Option<AssetMetadataValueDetail<T::Moment>>,
-    ) -> DispatchResult {
-        // Ensure the caller has the correct permissions for this asset.
-        let did = <ExternalAgents<T>>::ensure_perms(origin, ticker)?;
-
-        Self::unverified_set_asset_metadata(did, ticker, key, value, detail)
-    }
-
-    fn unverified_set_asset_metadata(
-        did: IdentityId,
-        ticker: Ticker,
-        key: AssetMetadataKey,
-        value: AssetMetadataValue,
-        detail: Option<AssetMetadataValueDetail<T::Moment>>,
-    ) -> DispatchResult {
-        // Check value length limit.
-        Self::ensure_asset_metadata_value_limited(&value)?;
-
-        // Check key exists.
-        ensure!(
-            Self::check_asset_metadata_key_exists(&ticker, &key),
-            Error::<T>::AssetMetadataKeyIsMissing
-        );
-
-        // Check if value is currently locked.
-        ensure!(
-            !Self::is_asset_metadata_locked(ticker, key),
-            Error::<T>::AssetMetadataValueIsLocked
-        );
-
-        // Set asset metadata value for asset.
-        AssetMetadataValues::insert(ticker, key, &value);
-
-        // Set asset metadata value details.
-        if let Some(ref detail) = detail {
-            AssetMetadataValueDetails::<T>::insert(ticker, key, detail);
-        }
-
-        Self::deposit_event(RawEvent::SetAssetMetadataValue(did, ticker, value, detail));
-        Ok(())
-    }
-
-    fn base_set_asset_metadata_details(
-        origin: T::RuntimeOrigin,
-        ticker: Ticker,
-        key: AssetMetadataKey,
-        detail: AssetMetadataValueDetail<T::Moment>,
-    ) -> DispatchResult {
-        // Ensure the caller has the correct permissions for this asset.
-        let did = <ExternalAgents<T>>::ensure_perms(origin, ticker)?;
-
-        // Check key exists.
-        ensure!(
-            Self::check_asset_metadata_key_exists(&ticker, &key),
-            Error::<T>::AssetMetadataKeyIsMissing
-        );
-
-        // Check if value is currently locked.
-        ensure!(
-            !Self::is_asset_metadata_locked(ticker, key),
-            Error::<T>::AssetMetadataValueIsLocked
-        );
-
-        // Prevent locking an asset metadata with no value
-        if detail.is_locked(<pallet_timestamp::Pallet<T>>::get()) {
-            AssetMetadataValues::try_get(&ticker, &key)
-                .map_err(|_| Error::<T>::AssetMetadataValueIsEmpty)?;
-        }
-
-        // Set asset metadata value details.
-        AssetMetadataValueDetails::<T>::insert(ticker, key, &detail);
-
-        Self::deposit_event(RawEvent::SetAssetMetadataValueDetails(did, ticker, detail));
-        Ok(())
-    }
-
-    fn base_register_and_set_local_asset_metadata(
-        origin: T::RuntimeOrigin,
-        ticker: Ticker,
-        name: AssetMetadataName,
-        spec: AssetMetadataSpec,
-        value: AssetMetadataValue,
-        detail: Option<AssetMetadataValueDetail<T::Moment>>,
-    ) -> DispatchResult {
-        // Ensure the caller has the correct permissions for this asset.
-        let did = <ExternalAgents<T>>::ensure_perms(origin, ticker)?;
-
-        // Register local metadata type.
-        let key = Self::unverified_register_asset_metadata_local_type(did, ticker, name, spec)?;
-
-        Self::unverified_set_asset_metadata(did, ticker, key, value, detail)
-    }
-
-    fn base_register_asset_metadata_local_type(
-        origin: T::RuntimeOrigin,
-        ticker: Ticker,
-        name: AssetMetadataName,
-        spec: AssetMetadataSpec,
-    ) -> DispatchResult {
-        // Ensure the caller has the correct permissions for this asset.
-        let did = <ExternalAgents<T>>::ensure_perms(origin, ticker)?;
-
-        Self::unverified_register_asset_metadata_local_type(did, ticker, name, spec).map(drop)
-    }
-
-    fn unverified_register_asset_metadata_local_type(
-        did: IdentityId,
-        ticker: Ticker,
-        name: AssetMetadataName,
-        spec: AssetMetadataSpec,
-    ) -> Result<AssetMetadataKey, DispatchError> {
-        Self::ensure_asset_metadata_name_limited(&name)?;
-        Self::ensure_asset_metadata_spec_limited(&spec)?;
-
-        // Check if key already exists.
-        ensure!(
-            !AssetMetadataLocalNameToKey::contains_key(ticker, &name),
-            Error::<T>::AssetMetadataLocalKeyAlreadyExists
-        );
-
-        // Next local key for asset.
-        let key = AssetMetadataNextLocalKey::try_mutate(ticker, try_next_pre::<T, _>)?;
-
-        // Store local key <-> name mapping.
-        AssetMetadataLocalNameToKey::insert(ticker, &name, key);
-        AssetMetadataLocalKeyToName::insert(ticker, key, &name);
-
-        // Store local specs.
-        AssetMetadataLocalSpecs::insert(ticker, key, &spec);
-
-        Self::deposit_event(RawEvent::RegisterAssetMetadataLocalType(
-            did, ticker, name, key, spec,
-        ));
-        Ok(key.into())
-    }
-
-    fn base_register_asset_metadata_global_type(
-        origin: T::RuntimeOrigin,
-        name: AssetMetadataName,
-        spec: AssetMetadataSpec,
-    ) -> DispatchResult {
-        Self::ensure_asset_metadata_name_limited(&name)?;
-        Self::ensure_asset_metadata_spec_limited(&spec)?;
-
-        // Only allow global metadata types to be registered by root.
-        ensure_root(origin)?;
-
-        // Check if key already exists.
-        ensure!(
-            !AssetMetadataGlobalNameToKey::contains_key(&name),
-            Error::<T>::AssetMetadataGlobalKeyAlreadyExists
-        );
-
-        // Next global key.
-        let key = AssetMetadataNextGlobalKey::try_mutate(try_next_pre::<T, _>)?;
-
-        // Store global key <-> name mapping.
-        AssetMetadataGlobalNameToKey::insert(&name, key);
-        AssetMetadataGlobalKeyToName::insert(key, &name);
-
-        // Store global specs.
-        AssetMetadataGlobalSpecs::insert(key, &spec);
-
-        Self::deposit_event(RawEvent::RegisterAssetMetadataGlobalType(name, key, spec));
-        Ok(())
-    }
-
-    fn base_controller_transfer(
-        origin: T::RuntimeOrigin,
-        ticker: Ticker,
-        value: Balance,
-        from_portfolio: PortfolioId,
-        weight_meter: &mut WeightMeter,
-    ) -> DispatchResult {
-        let to_portfolio = Self::ensure_origin_ticker_and_portfolio_permissions(
-            origin,
-            ticker,
-            PortfolioKind::Default,
-            false,
-        )?;
-
-        // Transfer `value` of ticker tokens from `investor_did` to controller
-        Self::unsafe_transfer(
-            from_portfolio,
-            to_portfolio,
-            &ticker,
-            value,
-            None,
-            None,
-            to_portfolio.did,
-            weight_meter,
-        )?;
-        Self::deposit_event(RawEvent::ControllerTransfer(
-            to_portfolio.did,
-            ticker,
-            from_portfolio,
-            value,
-        ));
-        Ok(())
-    }
-
-    pub fn unsafe_can_transfer_granular(
-        from_custodian: Option<IdentityId>,
-        from_portfolio: PortfolioId,
-        to_custodian: Option<IdentityId>,
-        to_portfolio: PortfolioId,
-        ticker: &Ticker,
-        value: Balance,
-        weight_meter: &mut WeightMeter,
-    ) -> Result<GranularCanTransferResult, DispatchError> {
-        let invalid_granularity = Self::invalid_granularity(ticker, value);
-        let self_transfer = Self::self_transfer(&from_portfolio, &to_portfolio);
-        let invalid_receiver_cdd = Self::invalid_cdd(to_portfolio.did);
-        let invalid_sender_cdd = Self::invalid_cdd(from_portfolio.did);
-        let receiver_custodian_error =
-            Self::custodian_error(to_portfolio, to_custodian.unwrap_or(to_portfolio.did));
-        let sender_custodian_error =
-            Self::custodian_error(from_portfolio, from_custodian.unwrap_or(from_portfolio.did));
-        let sender_insufficient_balance =
-            Self::insufficient_balance(&ticker, from_portfolio.did, value);
-        let portfolio_validity_result = <Portfolio<T>>::ensure_portfolio_transfer_validity_granular(
-            &from_portfolio,
-            &to_portfolio,
-            ticker,
-            value,
-        );
-        let asset_frozen = Self::frozen(ticker);
-        let transfer_condition_result = Self::transfer_condition_failures_granular(
-            &from_portfolio.did,
-            &to_portfolio.did,
-            ticker,
-            value,
-            weight_meter,
-        )?;
-        let compliance_result = T::ComplianceManager::verify_restriction_granular(
-            ticker,
-            Some(from_portfolio.did),
-            Some(to_portfolio.did),
-            weight_meter,
-        )?;
-
-        Ok(GranularCanTransferResult {
-            invalid_granularity,
-            self_transfer,
-            invalid_receiver_cdd,
-            invalid_sender_cdd,
-            receiver_custodian_error,
-            sender_custodian_error,
-            sender_insufficient_balance,
-            asset_frozen,
-            result: !invalid_granularity
-                && !self_transfer
-                && !invalid_receiver_cdd
-                && !invalid_sender_cdd
-                && !receiver_custodian_error
-                && !sender_custodian_error
-                && !sender_insufficient_balance
-                && portfolio_validity_result.result
-                && !asset_frozen
-                && transfer_condition_result.iter().all(|result| result.result)
-                && compliance_result.result,
-            transfer_condition_result,
-            compliance_result,
-            consumed_weight: Some(weight_meter.consumed()),
-            portfolio_validity_result,
-        })
-    }
-
-    fn invalid_granularity(ticker: &Ticker, value: Balance) -> bool {
-        !Self::check_granularity(&ticker, value)
-    }
-
-    fn self_transfer(from: &PortfolioId, to: &PortfolioId) -> bool {
-        from.did == to.did
-    }
-
-    fn invalid_cdd(did: IdentityId) -> bool {
-        !Identity::<T>::has_valid_cdd(did)
-    }
-
-    fn custodian_error(from: PortfolioId, custodian: IdentityId) -> bool {
-        Portfolio::<T>::ensure_portfolio_custody(from, custodian).is_err()
-    }
-
-    fn insufficient_balance(ticker: &Ticker, did: IdentityId, value: Balance) -> bool {
-        Self::balance_of(&ticker, did) < value
-    }
-
-    fn portfolio_failure(
-        from_portfolio: &PortfolioId,
-        to_portfolio: &PortfolioId,
-        ticker: &Ticker,
-        value: Balance,
-    ) -> bool {
-        Portfolio::<T>::ensure_portfolio_transfer_validity(
-            from_portfolio,
-            to_portfolio,
-            ticker,
-            value,
-        )
-        .is_err()
-    }
-
-    fn statistics_failures(
-        from_did: &IdentityId,
-        to_did: &IdentityId,
-        ticker: &Ticker,
-        value: Balance,
-        weight_meter: &mut WeightMeter,
-    ) -> bool {
-        let total_supply = Self::total_supply(*ticker);
-        Statistics::<T>::verify_transfer_restrictions(
-            ticker,
-            from_did,
-            to_did,
-            Self::balance_of(ticker, from_did),
-            Self::balance_of(ticker, to_did),
-            value,
-            total_supply,
-            weight_meter,
-        )
-        .is_err()
-    }
-
-    fn transfer_condition_failures_granular(
-        from_did: &IdentityId,
-        to_did: &IdentityId,
-        ticker: &Ticker,
-        value: Balance,
-        weight_meter: &mut WeightMeter,
-    ) -> Result<Vec<TransferConditionResult>, DispatchError> {
-        let total_supply = Self::total_supply(*ticker);
-        Statistics::<T>::get_transfer_restrictions_results(
-            ticker,
-            from_did,
-            to_did,
-            Self::balance_of(ticker, from_did),
-            Self::balance_of(ticker, to_did),
-            value,
-            total_supply,
-            weight_meter,
-        )
-    }
-
-    fn base_register_custom_asset_type(
-        origin: T::RuntimeOrigin,
-        ty: Vec<u8>,
-    ) -> Result<CustomAssetTypeId, DispatchError> {
-        let did = Identity::<T>::ensure_perms(origin)?;
-        Self::unsafe_register_custom_asset_type(did, ty)
-    }
-
-    fn unsafe_register_custom_asset_type(
-        did: IdentityId,
-        ty: Vec<u8>,
-    ) -> Result<CustomAssetTypeId, DispatchError> {
-        ensure_string_limited::<T>(&ty)?;
-
-        Ok(match CustomTypesInverse::try_get(&ty) {
-            Ok(id) => {
-                Self::deposit_event(Event::<T>::CustomAssetTypeExists(did, id, ty));
-                id
-            }
-            Err(()) => {
-                let id = CustomTypeIdSequence::try_mutate(try_next_pre::<T, _>)?;
-                CustomTypesInverse::insert(&ty, id);
-                CustomTypes::insert(id, ty.clone());
-                Self::deposit_event(Event::<T>::CustomAssetTypeRegistered(did, id, ty));
-                id
-            }
-        })
-    }
-
-    fn base_update_asset_type(
-        origin: T::RuntimeOrigin,
-        ticker: Ticker,
-        asset_type: AssetType,
-    ) -> DispatchResult {
-        Self::ensure_asset_exists(&ticker)?;
-        Self::ensure_asset_type_valid(asset_type)?;
-        let did = <ExternalAgents<T>>::ensure_perms(origin, ticker)?;
-        Tokens::try_mutate(&ticker, |token| -> DispatchResult {
-            let token = token.as_mut().ok_or(Error::<T>::NoSuchAsset)?;
-            // Ensures that both parameters are non fungible types or if both are fungible types.
-            ensure!(
-                token.asset_type.is_fungible() == asset_type.is_fungible(),
-                Error::<T>::IncompatibleAssetTypeUpdate
-            );
-            token.asset_type = asset_type;
-            Ok(())
-        })?;
-        Self::deposit_event(RawEvent::AssetTypeChanged(did, ticker, asset_type));
-        Ok(())
-    }
-
-    /// Returns `None` if there's no asset associated to the given ticker,
-    /// returns Some(true) if the asset exists and is of type `AssetType::NonFungible`, and returns Some(false) otherwise.
-    pub fn nft_asset(ticker: &Ticker) -> Option<bool> {
-        let token = Tokens::try_get(ticker).ok()?;
-        Some(token.asset_type.is_non_fungible())
-    }
-
-    fn base_remove_local_metadata_key(
-        origin: T::RuntimeOrigin,
-        ticker: Ticker,
-        local_key: AssetMetadataLocalKey,
-    ) -> DispatchResult {
-        // Verifies if the caller has the correct permissions for this asset
-        let caller_did = <ExternalAgents<T>>::ensure_perms(origin, ticker)?;
-        // Verifies if the key exists.
-        let name = AssetMetadataLocalKeyToName::try_get(ticker, &local_key)
-            .map_err(|_| Error::<T>::AssetMetadataKeyIsMissing)?;
-        // Verifies if the value is locked
-        let metadata_key = AssetMetadataKey::Local(local_key);
-        if let Some(value_detail) = AssetMetadataValueDetails::<T>::get(&ticker, &metadata_key) {
-            ensure!(
-                !value_detail.is_locked(<pallet_timestamp::Pallet<T>>::get()),
-                Error::<T>::AssetMetadataValueIsLocked
-            );
-        }
-        // Verifies if the key belongs to an NFT collection
-        ensure!(
-            !T::NFTFn::is_collection_key(&ticker, &metadata_key),
-            Error::<T>::AssetMetadataKeyBelongsToNFTCollection
-        );
-        // Remove key from storage
-        AssetMetadataValues::remove(&ticker, &metadata_key);
-        AssetMetadataValueDetails::<T>::remove(&ticker, &metadata_key);
-        AssetMetadataLocalNameToKey::remove(&ticker, &name);
-        AssetMetadataLocalKeyToName::remove(&ticker, &local_key);
-        AssetMetadataLocalSpecs::remove(&ticker, &local_key);
-        Self::deposit_event(RawEvent::LocalMetadataKeyDeleted(
-            caller_did, ticker, local_key,
-        ));
-        Ok(())
-    }
-
-    fn base_remove_metadata_value(
-        origin: T::RuntimeOrigin,
-        ticker: Ticker,
-        metadata_key: AssetMetadataKey,
-    ) -> DispatchResult {
-        // Verifies if the caller has the correct permissions for this asset
-        let caller_did = <ExternalAgents<T>>::ensure_perms(origin, ticker)?;
-        // Verifies if the key exists.
-        match metadata_key {
-            AssetMetadataKey::Global(global_key) => {
-                if !AssetMetadataGlobalKeyToName::contains_key(&global_key) {
-                    return Err(Error::<T>::AssetMetadataKeyIsMissing.into());
-                }
-            }
-            AssetMetadataKey::Local(local_key) => {
-                if !AssetMetadataLocalKeyToName::contains_key(ticker, &local_key) {
-                    return Err(Error::<T>::AssetMetadataKeyIsMissing.into());
-                }
-            }
-        }
-        // Verifies if the value is locked
-        if let Some(value_detail) = AssetMetadataValueDetails::<T>::get(&ticker, &metadata_key) {
-            ensure!(
-                !value_detail.is_locked(<pallet_timestamp::Pallet<T>>::get()),
-                Error::<T>::AssetMetadataValueIsLocked
-            );
-        }
-        // Remove the metadata value from storage
-        AssetMetadataValues::remove(&ticker, &metadata_key);
-        AssetMetadataValueDetails::<T>::remove(&ticker, &metadata_key);
-        Self::deposit_event(RawEvent::MetadataValueDeleted(
-            caller_did,
-            ticker,
-            metadata_key,
-        ));
-        Ok(())
-    }
-
-    /// Pre-approves the receivement of the asset for all identities.
-    fn base_exempt_ticker_affirmation(origin: T::RuntimeOrigin, ticker: Ticker) -> DispatchResult {
-        ensure_root(origin)?;
-        TickersExemptFromAffirmation::insert(&ticker, true);
-        Self::deposit_event(RawEvent::AssetAffirmationExemption(ticker));
-        Ok(())
-    }
-
-    /// Removes the pre-approval of the asset for all identities.
-    fn base_remove_ticker_affirmation_exemption(
-        origin: T::RuntimeOrigin,
-        ticker: Ticker,
-    ) -> DispatchResult {
-        ensure_root(origin)?;
-        TickersExemptFromAffirmation::remove(&ticker);
-        Self::deposit_event(RawEvent::RemoveAssetAffirmationExemption(ticker));
-        Ok(())
-    }
-
-    /// Pre-approves the receivement of an asset.
-    fn base_pre_approve_ticker(origin: T::RuntimeOrigin, ticker: Ticker) -> DispatchResult {
-        let caller_did = Identity::<T>::ensure_perms(origin)?;
-        PreApprovedTicker::insert(&caller_did, &ticker, true);
-        Self::deposit_event(RawEvent::PreApprovedAsset(caller_did, ticker));
-        Ok(())
-    }
-
-    /// Removes the pre approval of an asset.
-    fn base_remove_ticker_pre_approval(origin: T::RuntimeOrigin, ticker: Ticker) -> DispatchResult {
-        let caller_did = Identity::<T>::ensure_perms(origin)?;
-        PreApprovedTicker::remove(&caller_did, &ticker);
-        Self::deposit_event(RawEvent::RemovePreApprovedAsset(caller_did, ticker));
-        Ok(())
-    }
-
-    /// Sets all identities in the `mediators` set as mandatory mediators for any instruction transfering `ticker`.
-    fn base_add_mandatory_mediators(
-        origin: T::RuntimeOrigin,
-        ticker: Ticker,
-        new_mediators: BoundedBTreeSet<IdentityId, T::MaxAssetMediators>,
-    ) -> DispatchResult {
-        // Verifies if the caller has the correct permissions for this asset
-        let caller_did = <ExternalAgents<T>>::ensure_perms(origin, ticker)?;
-        // Tries to add all new identities as mandatory mediators for the asset
-        MandatoryMediators::<T>::try_mutate(ticker, |mandatory_mediators| -> DispatchResult {
-            for new_mediator in &new_mediators {
-                mandatory_mediators
-                    .try_insert(*new_mediator)
-                    .map_err(|_| Error::<T>::NumberOfAssetMediatorsExceeded)?;
-            }
-            Ok(())
-        })?;
-
-        Self::deposit_event(RawEvent::AssetMediatorsAdded(
-            caller_did,
-            ticker,
-            new_mediators.into_inner(),
-        ));
-        Ok(())
-    }
-
-    /// Removes all identities in the `mediators` set from the mandatory mediators list for the given `ticker`.
-    fn base_remove_mandatory_mediators(
-        origin: T::RuntimeOrigin,
-        ticker: Ticker,
-        mediators: BoundedBTreeSet<IdentityId, T::MaxAssetMediators>,
-    ) -> DispatchResult {
-        // Verifies if the caller has the correct permissions for this asset
-        let caller_did = <ExternalAgents<T>>::ensure_perms(origin, ticker)?;
-        // Removes the identities from the mandatory mediators list
-        MandatoryMediators::<T>::mutate(ticker, |mandatory_mediators| {
-            for mediator in &mediators {
-                mandatory_mediators.remove(mediator);
-            }
-        });
-        Self::deposit_event(RawEvent::AssetMediatorsRemoved(
-            caller_did,
-            ticker,
-            mediators.into_inner(),
-        ));
-        Ok(())
-    }
-}
-
-impl<T: Config> AssetFnTrait<T::AccountId, T::RuntimeOrigin> for Module<T> {
-    fn ensure_granular(ticker: &Ticker, value: Balance) -> DispatchResult {
-        Self::ensure_granular(ticker, value)
-    }
-
-    /// Get the asset `id` balance of `who`.
-    fn balance(ticker: &Ticker, who: IdentityId) -> Balance {
-        Self::balance_of(ticker, &who)
-    }
-
-    fn create_asset(
-        origin: T::RuntimeOrigin,
-        name: AssetName,
-        ticker: Ticker,
-        divisible: bool,
-        asset_type: AssetType,
-        identifiers: Vec<AssetIdentifier>,
-        funding_round: Option<FundingRoundName>,
-    ) -> DispatchResult {
-        Self::create_asset(
-            origin,
-            name,
-            ticker,
-            divisible,
-            asset_type,
-            identifiers,
-            funding_round,
-        )
-    }
-
-    fn register_ticker(origin: T::RuntimeOrigin, ticker: Ticker) -> DispatchResult {
-        Self::base_register_ticker(origin, ticker)
-    }
-
-    fn issue(
-        origin: T::RuntimeOrigin,
-        ticker: Ticker,
-        amount: Balance,
-        portfolio_kind: PortfolioKind,
-    ) -> DispatchResult {
-        Self::issue(origin, ticker, amount, portfolio_kind)
-    }
-
-    fn skip_ticker_affirmation(identity_id: &IdentityId, ticker: &Ticker) -> bool {
-        if TickersExemptFromAffirmation::get(ticker) {
-            return true;
-        }
-        PreApprovedTicker::get(identity_id, ticker)
-    }
-
-    fn ticker_affirmation_exemption(ticker: &Ticker) -> bool {
-        TickersExemptFromAffirmation::get(ticker)
-    }
-
-    #[cfg(feature = "runtime-benchmarks")]
-    fn register_asset_metadata_type(
-        origin: T::RuntimeOrigin,
-        ticker: Option<Ticker>,
-        name: AssetMetadataName,
-        spec: AssetMetadataSpec,
-    ) -> DispatchResult {
-        match ticker {
-            Some(ticker) => Self::register_asset_metadata_local_type(origin, ticker, name, spec),
-            None => Self::register_asset_metadata_global_type(origin, name, spec),
-        }
-    }
-
-    #[cfg(feature = "runtime-benchmarks")]
-    fn add_mandatory_mediators(
-        origin: T::RuntimeOrigin,
-        ticker: Ticker,
-        mediators: BTreeSet<IdentityId>,
-    ) -> DispatchResult {
-        Self::add_mandatory_mediators(origin, ticker, mediators.try_into().unwrap_or_default())
     }
 }
