+++ pallets/runtime/tests/src/corporate_actions_test.rs
@@ -2112,7 +2112,7 @@ fn dist_claim_works() {
         // Owner should have some free currency balance due to withheld taxes.
         let pid = PortfolioId::default_portfolio(owner.did);
         let wht = benefit_foo - post_tax_foo + benefit_bar - post_tax_bar;
-        let rem = Asset::total_supply(ticker) - amount + wht;
+        let rem = Asset::total_supply(&ticker) - amount + wht;
         assert_ok!(Portfolio::ensure_sufficient_balance(&pid, &currency, rem));
         assert_noop!(
             Portfolio::ensure_sufficient_balance(&pid, &currency, rem + 1),
